Update to upstream version '3.0.1'
with Debian dir a4ecb845d5cceb7daa99c872cb257eaeb3d1d361

+++ setup/setup.sh
@@ -38,38 +38,42 @@ case ${osinfo} in
   # Kali 2 dependency Install
   Kali2)
     apt-get update
-    echo '[*] Installing Kali2 Dependencies'
-    apt-get install -y cmake qt4-qmake python xvfb python-qt4 python-pip python-netaddr python-dev tesseract-ocr
+    echo '[*] Installing Debian Dependencies'
+    apt-get install -y cmake python3 xvfb python3-pip python-netaddr python3-dev tesseract-ocr firefox-esr
     echo '[*] Upgrading paramiko'
-    pip install --upgrade paramiko
-    echo '[*] Installing RDPY'
-    git clone https://github.com/ChrisTruncer/rdpy.git
-    cd rdpy
-    python setup.py install
-    cd ..
-    rm -rf rdpy
+    python3 -m pip install --upgrade paramiko
+    echo
     echo '[*] Installing Python Modules'
-    pip install fuzzywuzzy
-    pip install selenium --upgrade
-    pip install python-Levenshtein
-    pip install pyasn1 --upgrade
-    pip install pyvirtualdisplay
-    pip install pytesseract
+    python3 -m pip install fuzzywuzzy
+    python3 -m pip install selenium --upgrade
+    python3 -m pip install python-Levenshtein
+    python3 -m pip install pyasn1
+    python3 -m pip install pyvirtualdisplay
+    python3 -m pip install beautifulsoup4
+    python3 -m pip install pytesseract
+    python3 -m pip install netaddr
+    echo
     cd ../bin/
     MACHINE_TYPE=`uname -m`
     if [ ${MACHINE_TYPE} == 'x86_64' ]; then
-      wget https://github.com/mozilla/geckodriver/releases/download/v0.22.0/geckodriver-v0.22.0-linux64.tar.gz
-      tar -xvf geckodriver-v0.22.0-linux64.tar.gz
-      rm geckodriver-v0.22.0-linux64.tar.gz
+      wget https://github.com/mozilla/geckodriver/releases/download/v0.24.0/geckodriver-v0.24.0-linux64.tar.gz
+      tar -xvf geckodriver-v0.24.0-linux64.tar.gz
+      rm geckodriver-v0.24.0-linux64.tar.gz
       mv geckodriver /usr/sbin
-      rm /usr/bin/geckodriver
+      if [ -e /usr/bin/geckodriver ]
+      then
+      	rm /usr/bin/geckodriver
+      fi
       ln -s /usr/sbin/geckodriver /usr/bin/geckodriver
     else
-      wget https://github.com/mozilla/geckodriver/releases/download/v0.22.0/geckodriver-v0.22.0-linux32.tar.gz
-      tar -xvf geckodriver-v0.22.0-linux32.tar.gz
-      rm geckodriver-v0.22.0-linux64.tar.gz
+      wget https://github.com/mozilla/geckodriver/releases/download/v0.24.0/geckodriver-v0.24.0-linux32.tar.gz
+      tar -xvf geckodriver-v0.24.0-linux32.tar.gz
+      rm geckodriver-v0.24.0-linux32.tar.gz
       mv geckodriver /usr/sbin
-      rm /usr/bin/geckodriver
+      if [ -e /usr/bin/geckodriver ]
+      then
+      	rm /usr/bin/geckodriver
+      fi
       ln -s /usr/sbin/geckodriver /usr/bin/geckodriver
     fi
     cd ..
@@ -77,88 +81,128 @@ case ${osinfo} in
   # Kali Dependency Installation
   Kali)
     apt-get update
-    echo '[*] Installing Kali Dependencies'
-    apt-get install -y python-qt4 python-pip xvfb python-netaddr python-dev tesseract-ocr
+    echo '[*] Installing Debian Dependencies'
+    apt-get install -y cmake python3 xvfb python3-pip python-netaddr python3-dev tesseract-ocr firefox-esr
     echo '[*] Upgrading paramiko'
-    pip install --upgrade paramiko
-    echo '[*] Installing RDPY'
-    git clone https://github.com/ChrisTruncer/rdpy.git
-    cd rdpy
-    python setup.py install
-    cd ..
-    rm -rf rdpy
+    python3 -m pip install --upgrade paramiko
+    echo
     echo '[*] Installing Python Modules'
-    pip install fuzzywuzzy
-    pip install selenium --upgrade
-    pip install python-Levenshtein
-    pip install pyasn1 --upgrade
-    pip install pyvirtualdisplay
-    pip install pytesseract
+    python3 -m pip install fuzzywuzzy
+    python3 -m pip install selenium --upgrade
+    python3 -m pip install python-Levenshtein
+    python3 -m pip install pyasn1
+    python3 -m pip install pyvirtualdisplay
+    python3 -m pip install beautifulsoup4
+    python3 -m pip install pytesseract
+    python3 -m pip install netaddr
+    echo
     cd ../bin/
+    MACHINE_TYPE=`uname -m`
+    if [ ${MACHINE_TYPE} == 'x86_64' ]; then
+      wget https://github.com/mozilla/geckodriver/releases/download/v0.24.0/geckodriver-v0.24.0-linux64.tar.gz
+      tar -xvf geckodriver-v0.24.0-linux64.tar.gz
+      rm geckodriver-v0.24.0-linux64.tar.gz
+      mv geckodriver /usr/sbin
+      if [ -e /usr/bin/geckodriver ]
+      then
+      	rm /usr/bin/geckodriver
+      fi
+      ln -s /usr/sbin/geckodriver /usr/bin/geckodriver
+    else
+      wget https://github.com/mozilla/geckodriver/releases/download/v0.24.0/geckodriver-v0.24.0-linux32.tar.gz
+      tar -xvf geckodriver-v0.24.0-linux32.tar.gz
+      rm geckodriver-v0.24.0-linux32.tar.gz
+      mv geckodriver /usr/sbin
+      if [ -e /usr/bin/geckodriver ]
+      then
+      	rm /usr/bin/geckodriver
+      fi
+      ln -s /usr/sbin/geckodriver /usr/bin/geckodriver
+    fi
     cd ..
   ;;
    # Parrot Dependency Installation
   Parrot)
     apt-get update
-    echo '[*] Installing Parrot Dependencies'
-    apt-get install -y python-qt4 python-pip xvfb python-netaddr python-dev tesseract-ocr firefox-esr
+    echo '[*] Installing Debian Dependencies'
+    apt-get install -y cmake python3 xvfb python3-pip python-netaddr python3-dev tesseract-ocr firefox-esr
     echo '[*] Upgrading paramiko'
-    pip install --upgrade paramiko
-    echo '[*] Installing RDPY'
-    git clone https://github.com/ChrisTruncer/rdpy.git
-    cd rdpy
-    python setup.py install
-    cd ..
-    rm -rf rdpy
+    python3 -m pip install --upgrade paramiko
+    echo
     echo '[*] Installing Python Modules'
-    pip install fuzzywuzzy
-    pip install selenium --upgrade
-    pip install python-Levenshtein
-    pip install pyasn1 --upgrade
-    pip install pyvirtualdisplay
-    pip install pytesseract
+    python3 -m pip install fuzzywuzzy
+    python3 -m pip install selenium --upgrade
+    python3 -m pip install python-Levenshtein
+    python3 -m pip install pyasn1
+    python3 -m pip install pyvirtualdisplay
+    python3 -m pip install beautifulsoup4
+    python3 -m pip install pytesseract
+    python3 -m pip install netaddr
+    echo
     cd ../bin/
+    MACHINE_TYPE=`uname -m`
+    if [ ${MACHINE_TYPE} == 'x86_64' ]; then
+      wget https://github.com/mozilla/geckodriver/releases/download/v0.24.0/geckodriver-v0.24.0-linux64.tar.gz
+      tar -xvf geckodriver-v0.24.0-linux64.tar.gz
+      rm geckodriver-v0.24.0-linux64.tar.gz
+      mv geckodriver /usr/sbin
+      if [ -e /usr/bin/geckodriver ]
+      then
+      	rm /usr/bin/geckodriver
+      fi
+      ln -s /usr/sbin/geckodriver /usr/bin/geckodriver
+    else
+      wget https://github.com/mozilla/geckodriver/releases/download/v0.24.0/geckodriver-v0.24.0-linux32.tar.gz
+      tar -xvf geckodriver-v0.24.0-linux32.tar.gz
+      rm geckodriver-v0.24.0-linux32.tar.gz
+      mv geckodriver /usr/sbin
+      if [ -e /usr/bin/geckodriver ]
+      then
+      	rm /usr/bin/geckodriver
+      fi
+      ln -s /usr/sbin/geckodriver /usr/bin/geckodriver
+    fi
     cd ..
   ;;
   # Debian 7+ Dependency Installation
   Debian)
     apt-get update
     echo '[*] Installing Debian Dependencies'
-    apt-get install -y cmake qt4-qmake python xvfb python-qt4 python-pip python-netaddr python-dev tesseract-ocr firefox-esr
+    apt-get install -y cmake python3 xvfb python3-pip python-netaddr python3-dev tesseract-ocr firefox-esr
     echo '[*] Upgrading paramiko'
-    pip install --upgrade paramiko
-    echo '[*] Installing RDPY'
-    git clone https://github.com/ChrisTruncer/rdpy.git
-    cd rdpy
-    python setup.py install
-    cd ..
-    rm -rf rdpy
+    python3 -m pip install --upgrade paramiko
     echo
     echo '[*] Installing Python Modules'
-    pip install python_qt_binding
-    pip install fuzzywuzzy
-    pip install selenium --upgrade
-    pip install python-Levenshtein
-    pip install pyasn1
-    pip install pyvirtualdisplay
-    pip install beautifulsoup4
-    pip install pytesseract
+    python3 -m pip install fuzzywuzzy
+    python3 -m pip install selenium --upgrade
+    python3 -m pip install python-Levenshtein
+    python3 -m pip install pyasn1
+    python3 -m pip install pyvirtualdisplay
+    python3 -m pip install beautifulsoup4
+    python3 -m pip install pytesseract
+    python3 -m pip install netaddr
     echo
     cd ../bin/
     MACHINE_TYPE=`uname -m`
     if [ ${MACHINE_TYPE} == 'x86_64' ]; then
-      wget https://github.com/mozilla/geckodriver/releases/download/v0.22.0/geckodriver-v0.22.0-linux64.tar.gz
-      tar -xvf geckodriver-v0.22.0-linux64.tar.gz
-      rm geckodriver-v0.22.0-linux64.tar.gz
+      wget https://github.com/mozilla/geckodriver/releases/download/v0.24.0/geckodriver-v0.24.0-linux64.tar.gz
+      tar -xvf geckodriver-v0.24.0-linux64.tar.gz
+      rm geckodriver-v0.24.0-linux64.tar.gz
       mv geckodriver /usr/sbin
-      rm /usr/bin/geckodriver
+      if [ -e /usr/bin/geckodriver ]
+      then
+      	rm /usr/bin/geckodriver
+      fi
       ln -s /usr/sbin/geckodriver /usr/bin/geckodriver
     else
-      wget https://github.com/mozilla/geckodriver/releases/download/v0.22.0/geckodriver-v0.22.0-linux32.tar.gz
-      tar -xvf geckodriver-v0.22.0-linux32.tar.gz
-      rm geckodriver-v0.22.0-linux64.tar.gz
+      wget https://github.com/mozilla/geckodriver/releases/download/v0.24.0/geckodriver-v0.24.0-linux32.tar.gz
+      tar -xvf geckodriver-v0.24.0-linux32.tar.gz
+      rm geckodriver-v0.24.0-linux32.tar.gz
       mv geckodriver /usr/sbin
-      rm /usr/bin/geckodriver
+      if [ -e /usr/bin/geckodriver ]
+      then
+      	rm /usr/bin/geckodriver
+      fi
       ln -s /usr/sbin/geckodriver /usr/bin/geckodriver
     fi
     cd ..
@@ -166,75 +210,44 @@ case ${osinfo} in
   # Ubuntu (tested in 13.10) Dependency Installation
   Ubuntu)
     apt-get update
-    echo '[*] Installing Ubuntu Dependencies'
-    apt-get install -y cmake qt4-qmake python python-qt4 python-pip xvfb python-netaddr python-dev libffi-dev libssl-dev tesseract-ocr firefox-esr
-    echo '[*] Installing RDPY'
-    git clone https://github.com/ChrisTruncer/rdpy.git
-    cd rdpy
-    python setup.py install
-    cd ..
-    rm -rf rdpy
+    echo '[*] Installing Debian Dependencies'
+    apt-get install -y cmake python3 xvfb python3-pip python-netaddr python3-dev tesseract-ocr firefox-esr
+    echo '[*] Upgrading paramiko'
+    python3 -m pip install --upgrade paramiko
     echo
     echo '[*] Installing Python Modules'
-    pip install python_qt_binding
-    pip install fuzzywuzzy
-    pip install selenium --upgrade
-    pip install python-Levenshtein
-    pip install pyasn1
-    pip install pyvirtualdisplay
-    pip install beautifulsoup4
-    pip install pytesseract
-    pip install enum34
-    pip install ipaddress
-    pip install asn1crypto
+    python3 -m pip install fuzzywuzzy
+    python3 -m pip install selenium --upgrade
+    python3 -m pip install python-Levenshtein
+    python3 -m pip install pyasn1
+    python3 -m pip install pyvirtualdisplay
+    python3 -m pip install beautifulsoup4
+    python3 -m pip install pytesseract
+    python3 -m pip install netaddr
     echo
     cd ../bin/
     MACHINE_TYPE=`uname -m`
     if [ ${MACHINE_TYPE} == 'x86_64' ]; then
-      wget https://github.com/mozilla/geckodriver/releases/download/v0.22.0/geckodriver-v0.22.0-linux64.tar.gz
-      tar -xvf geckodriver-v0.22.0-linux64.tar.gz
-      rm geckodriver-v0.22.0-linux64.tar.gz
+      wget https://github.com/mozilla/geckodriver/releases/download/v0.24.0/geckodriver-v0.24.0-linux64.tar.gz
+      tar -xvf geckodriver-v0.24.0-linux64.tar.gz
+      rm geckodriver-v0.24.0-linux64.tar.gz
       mv geckodriver /usr/sbin
+      if [ -e /usr/bin/geckodriver ]
+      then
+      	rm /usr/bin/geckodriver
+      fi
+      ln -s /usr/sbin/geckodriver /usr/bin/geckodriver
     else
-      wget https://github.com/mozilla/geckodriver/releases/download/v0.22.0/geckodriver-v0.22.0-linux32.tar.gz
-      tar -xvf geckodriver-v0.22.0-linux32.tar.gz
-      rm geckodriver-v0.22.0-linux64.tar.gz
+      wget https://github.com/mozilla/geckodriver/releases/download/v0.24.0/geckodriver-v0.24.0-linux32.tar.gz
+      tar -xvf geckodriver-v0.24.0-linux32.tar.gz
+      rm geckodriver-v0.24.0-linux32.tar.gz
       mv geckodriver /usr/sbin
+      if [ -e /usr/bin/geckodriver ]
+      then
+      	rm /usr/bin/geckodriver
+      fi
+      ln -s /usr/sbin/geckodriver /usr/bin/geckodriver
     fi
-    cd ..
-  ;;
-  # CentOS 6.5+ Dependency Installation
-  CentOS)
-    echo '[Warning]: EyeWitness on CentOS Requires EPEL Repository!'
-    read -p '[?] Install and Enable EPEL Repository? (y/n): ' epel
-    if [ "${epel}" == 'y' ]; then
-      rpm -ivh ${eplpkg}
-    else
-      echo '[!] User Aborted EyeWitness Installation.'
-      popd > /dev/null
-      exit 1
-    fi
-    echo '[*] Installing CentOS Dependencies'
-    yum install cmake python python-pip PyQt4 PyQt4-webkit \
-                python-argparse xvfb python-netaddr python-dev tesseract-ocr firefox-esr
-    echo
-    echo '[*] Installing RDPY'
-    git clone https://github.com/ChrisTruncer/rdpy.git
-    cd rdpy
-    python setup.py install
-    cd ..
-    rm -rf rdpy
-    echo '[*] Installing Python Modules'
-    pip install python_qt_binding
-    pip install fuzzywuzzy
-    pip install selenium --upgrade
-    pip install python-Levenshtein
-    pip install pyasn1
-    pip install pyvirtualdisplay
-    pip install beautifulsoup4
-    pip install pytesseract
-    echo
-    cd ../bin/
     cd ..
   ;;
   # Notify Manual Installation Requirement And Exit
